refactor: move validations to crypto
