+++ packages/crypto/lib/validation/extensions/index.js
