+++ packages/crypto/lib/validation/extensions/public-key.js
