+++ packages/crypto/lib/validation/index.js
@@ -1,5 +1,3 @@
-'use strict'
-
 module.exports = {
   validator: require('./validator'),
   transactionValidator: require('./validators/transaction')
