+++ packages/crypto/lib/validation/rules/models/transactions/transfer.js
@@ -1,24 +1,20 @@
-const { TRANSACTION_TYPES } = require('../../../constants')
+const { TRANSACTION_TYPES } = require('../../../../constants')
 const engine = require('../../../engine')
 
 module.exports = (transaction) => {
   const { error, value } = engine.validate(transaction, engine.joi.object({
     id: engine.joi.string().alphanum().required(),
     blockid: engine.joi.number(),
-    type: engine.joi.number().valid(TRANSACTION_TYPES.SECOND_SIGNATURE),
+    type: engine.joi.number().valid(TRANSACTION_TYPES.TRANSFER),
     timestamp: engine.joi.number().min(0).required(),
-    amount: engine.joi.number().valid(0),
+    amount: engine.joi.number().min(1).required(),
     fee: engine.joi.number().min(1).required(),
     senderId: engine.joi.arkAddress(),
+    recipientId: engine.joi.arkAddress().required(),
     senderPublicKey: engine.joi.arkPublicKey().required(),
     signature: engine.joi.string().alphanum().required(),
     signatures: engine.joi.array(),
-    secondSignature: engine.joi.empty(),
-    asset: engine.joi.object({
-      signature: engine.joi.object({
-        publicKey: engine.joi.arkPublicKey().required()
-      }).required()
-    }).required(),
+    secondSignature: engine.joi.string().alphanum(),
     confirmations: engine.joi.number().min(0)
   }), {
     allowUnknown: true
