+++ packages/crypto/lib/validation/rules/public-key.js
