+++ packages/crypto/package.json
@@ -30,7 +30,6 @@
     "depcheck": "depcheck ./"
   },
   "dependencies": {
-    "@arkecosystem/validation": "^0.1.1",
     "bigi": "^1.4.2",
     "bip38": "^2.0.2",
     "bip66": "^1.1.5",
@@ -40,6 +39,7 @@
     "create-hmac": "^1.1.7",
     "deepmerge": "^2.1.0",
     "ecurve": "^1.0.6",
+    "joi": "^13.3.0",
     "lodash": "^4.17.10",
     "moment": "^2.22.1",
     "node-forge": "^0.7.5",
